changing the bias correction choice for TwoTowerDNN to be confiured at the tuning stage.  Because the TwoTowerDNN is no longer a regression model, ratings_train_liked.dat was created without the ratings < 4 rows.

+++ src/main/python/movie_lens_tfx/transform_movie_lens_metadata.py
@@ -44,7 +44,7 @@ def preprocessing_fn(inputs):
   'user_id': <tf.Tensor 'inputs_7_copy:0' shape=(None, 1) dtype=int64>}
     
     outputs={
-    'movie_id': <tf.Tensor 'Cast_1:0' shape=(None, 1) dtype=float32>,
+    'movie_id': <tf.Tensor 'Cast_1:0' shape=(None, 1) dtype=int32>,
     'rating': <tf.Tensor 'truediv:0' shape=(None, 1) dtype=float32>,
     'genres': <tf.Tensor 'RaggedToTensor_1/RaggedTensorToTensor:0' shape=(None, 1, 18) dtype=float32>,
       }
@@ -62,6 +62,7 @@ def preprocessing_fn(inputs):
         return tf.lookup.StaticHashTable(init, default_value=-1)
     genres_table = create_static_table(genres, var_dtype=tf.string)
     
+  #TODO: change the integer types to int32
   outputs = {'movie_id': tf.cast(inputs['movie_id'], dtype=tf.float32)}
 
   outputs['rating'] = tf.divide(
